Fixed maths functions

+++ esp32/main/avoid.c
@@ -4,15 +4,15 @@
 static bool isInEllipse(vect_2d_t robot, float a, float b)
 {
     vect_2d_t temp1 = vect_2d(robot.x, robot.y, false);
-    float angle = atan2(robot.y, robot.x);
+    float angle = atan2f(robot.y, robot.x);
     vect_2d_t temp2 = vect_2d(a * cos(angle), b * sin(angle), false);
     return temp2.mag > temp1.mag;
 }
 static vect_2d_t ellipsePoint(vect_2d_t robot, float a, float b)
 {
     //replace with better nearest point function when angus is done
-    float angle = atan2(robot.y, robot.x);
-    return vect_2d(a * cos(angle), b * sin(angle), false);
+    float angle = atan2f(robot.y, robot.x);
+    return vect_2d(a * cosf(angle), b * sinf(angle), false);
 }
 
 static inline float mCalc(vect_2d_t robot, float a, float b)
@@ -29,14 +29,14 @@ static inline float x(vect_2d_t robot, float a, float b, int p)
 {
     float m = mCalc(robot, a, b);
     float c = cCalc(robot, a, b);
-    return ((-powf(a, 2) * m * c) + p * sign(robot.y) * a * b * sqrtf((pow(a, 2) * pow(m, 2)) + pow(b, 2) - pow(c, 2))) / (pow(b, 2) + (pow(a, 2) * pow(m, 2)));
+    return ((-powf(a, 2) * m * c) + p * sign(robot.y) * a * b * sqrtf((powf(a, 2) * powf(m, 2)) + powf(b, 2) - powf(c, 2))) / (powf(b, 2) + (powf(a, 2) * powf(m, 2)));
 }
 
 static inline float y(vect_2d_t robot, float a, float b, int p)
 {
     float m = mCalc(robot, a, b);
     float c = cCalc(robot, a, b);
-    return ((powf(b, 2) * c) + p * sign(robot.y) * a * b * m * sqrtf((pow(a, 2) * pow(m, 2)) + pow(b, 2) - pow(c, 2))) / (pow(b, 2) + (pow(a, 2) * pow(m, 2)));
+    return ((powf(b, 2) * c) + p * sign(robot.y) * a * b * m * sqrtf((powf(a, 2) * powf(m, 2)) + powf(b, 2) - powf(c, 2))) / (powf(b, 2) + (powf(a, 2) * powf(m, 2)));
 }
 
 static vect_2d_t calcAvoid(float a, float b, vect_2d_t obj, vect_2d_t robot)
